Move map code to new view file
